+++ projects/services-layers/src/lib/types/Layers.ts
@@ -59,6 +59,13 @@ export function isLayertype(inpt: string): inpt is TLayertype {
  */
 export type TGeoExtent = [number, number, number, number] | [number, number, number, number, number, number];
 
+/*
+ * There are effectively only two values that we may set for cors:
+ * https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/crossorigin
+ */
+export type CrossOriginType = 'anonymous' | 'use-credentials';
+
+
 export interface ILayerOptions {
   name: string;
   id: string;
@@ -87,6 +94,8 @@ export interface ILayerOptions {
   actions?: [{ title: string, icon: string, action: (Layer) => void }];
   /** a layer might have more than one style; eg. true color and false color for the same dataset */
   styles?: ILayerStyleSet[];
+  /** The crossOrigin attribute for loaded images if you want to access pixel data with the Canvas renderer */
+  crossOrigin?: CrossOriginType;
 }
 
 export interface ILayerDimensions extends IAnyObject {
@@ -116,6 +125,9 @@ export interface ILayerElevationDimension {
   value?: string;
 }
 
+
+
+
 export interface IRasterLayerOptions extends ILayerOptions {
   url: string;
   subdomains?: Array<string>;
@@ -142,7 +154,12 @@ export interface IVectorLayerOptions extends ILayerOptions {
   type: TVectorLayertype;
 }
 
-export interface ICustomLayerOptions extends Omit<ILayerOptions, 'type'> {
+/**
+ * Deliberately does not have the `crossOrigin` property.
+ * When using openlayers as a map-engine, `crossOrigin` is a property that UKIS just passes on to the layersource.
+ * Since that layersource is provided by the user in a CustomLayer, setting `crossOrigin` in the ICustomLayerOptions would have no effect.
+ */
+export interface ICustomLayerOptions extends Omit<ILayerOptions, 'type' | 'crossOrigin'> {
   type?: TLayertype;
   custom_layer: any;
 }
@@ -178,7 +195,7 @@ export class Layer implements ILayerOptions {
   /** a layer might have more than one style; eg. true color and false color for the same dataset */
   styles?: ILayerStyleSet[];
   /** The crossOrigin attribute for loaded images if you want to access pixel data with the Canvas renderer */
-  crossOrigin?: string;
+  crossOrigin?: CrossOriginType;
 
   constructor(options: ILayerOptions) {
     Object.assign(this, options);